Refactor pausable field plugin to have common codebase (#118909)
(cherry picked from commit 65faabd08d79043d10a0351e57f1b0c6239862da)
